Fix enemy can be targeted by multiple towers
- Renamed turret to tower
